ecc-openlxp-xis/requirements.txt
bleach~=6.0.0

celery >=5.2.3, <6.0

confusable_homoglyphs==3.2.0

coverage>=5.5,<6.0

ddt~=1.4.2 , <1.5.0

django-admin-interface>=0.28.3

django-celery-beat >=2.5.0, <2.6.0

django-celery-results~=2.4.0

django-colorfield>=0.11.0

django-cors-headers>=3.7.0, <3.8.0

django-csp>=3.7

django-model-utils>=4.3.1,<5.0.0

django-mysql>=4.4.0,<5.0.0

django-rest-knox>=5.0.2,<5.1.0

django-ses>=2.2.0 , <2.3.1

Django>=4.2.24,<5.0

djangorestframework>=3.15.2,<3.16.0

elasticsearch==7.11.0

flake8>=3.9.0,<4.0.1

gunicorn>=23.0.0,<23.1.0

mysql-connector-python >= 9.1.0, <9.2.0

neo4j >=4.4.1

numpy==1.26.4

openlxp-P1-notification >=0.0.18, <2.0.0

p1-auth>=0.2.4,<1.0.0

pandas>=1.3.5,<1.4.0

Pillow>=11.3.0, <11.4.0

py2neo >=2021.2.3

pylint-django == 2.5.3

redis==6.4.0

requests~=2.32.4

setuptools==80.9.0